+++ plexsvc-framework/src/main/java/com/plexobject/encode/CodecConfigurer.java
@@ -1,0 +1,11 @@
+package com.plexobject.encode;
+
+/**
+ * This interface allows configuration of underlying encoder
+ * 
+ * @author shahzad bhatti
+ *
+ */
+public interface CodecConfigurer {
+    void configureCodec(Object underlyingEncoder);
+}